reward takes the number of steps survived into account

+++ flappy/lib_flappy/flappy.h
@@ -4,7 +4,7 @@ void reset_flappy(void);
 char* init_flappy(int display);
 void exit_flappy(void);
 void run_flappy(void);
-int step_flappy(int movement, int *reward);
+int step_flappy(int movement, float *reward);
 void treatingImage(char *image);
 void getSize(int *x_size, int *y_size);
 void updateFeatures(int *xToPipe, float *yToUpperPipe, float *yToLowerPipe, float *vy, float *yToTop, float *yToBottom);
